+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## Unrealesed
+## Added
+- Error definition
+## Changed
+- *Breaking change*: method signature `evaluate_model(&mut self)` to `evaluate_model(&mut self) -> Result<(), model::ModelError>`
+- *Breaking change*: method signature `get_jacobian(&self) -> residuals::JacobianValues<D>` to `get_jacobian(&mut self) -> Result<residuals::JacobianValues<D>, model::ModelError>` in the model trait.
 ## [0.8.0] - 2021-07-25
 ### Changed
 - Upgrade to nalgebra 0.28